(svn r2021) Fix: Enlarge _cur_town_ctr from byte to uint32 so that all the towns are considered when growing them even for big maps, where much more than 256 towns are around; reported by Tomasz Dubiński <uboottd@hydra.polsl.gliwice.pl>. The savegame still saves just the lowest 8 bits but that doesn't hurt so much.

+++ variables.h
@@ -84,7 +84,7 @@ VARDEF uint32 _random_seeds[2][2];
 VARDEF uint32 _player_seeds[MAX_PLAYERS][2];
 
 // Iterator through all towns in OnTick_Town
-VARDEF byte _cur_town_ctr;
+VARDEF uint32 _cur_town_ctr;
 
 VARDEF uint _cur_player_tick_index;
 VARDEF uint _next_competitor_start;
